+++ cli/src/commit_templater.rs
@@ -879,14 +879,7 @@ fn builtin_commit_methods<'repo>() -> CommitTemplateBuildMethodFnMap<'repo, Comm
     );
     map.insert(
         "bookmarks",
-        |language, diagnostics, _build_ctx, self_property, function| {
-            if function.name != "bookmarks" {
-                // TODO: Remove in jj 0.28+
-                diagnostics.add_warning(TemplateParseError::expression(
-                    "branches() is deprecated; use bookmarks() instead",
-                    function.name_span,
-                ));
-            }
+        |language, _diagnostics, _build_ctx, self_property, function| {
             function.expect_no_arguments()?;
             let index = language
                 .keyword_cache
@@ -905,14 +898,7 @@ fn builtin_commit_methods<'repo>() -> CommitTemplateBuildMethodFnMap<'repo, Comm
     );
     map.insert(
         "local_bookmarks",
-        |language, diagnostics, _build_ctx, self_property, function| {
-            if function.name != "local_bookmarks" {
-                // TODO: Remove in jj 0.28+
-                diagnostics.add_warning(TemplateParseError::expression(
-                    "local_branches() is deprecated; use local_bookmarks() instead",
-                    function.name_span,
-                ));
-            }
+        |language, _diagnostics, _build_ctx, self_property, function| {
             function.expect_no_arguments()?;
             let index = language
                 .keyword_cache
@@ -931,14 +917,7 @@ fn builtin_commit_methods<'repo>() -> CommitTemplateBuildMethodFnMap<'repo, Comm
     );
     map.insert(
         "remote_bookmarks",
-        |language, diagnostics, _build_ctx, self_property, function| {
-            if function.name != "remote_bookmarks" {
-                // TODO: Remove in jj 0.28+
-                diagnostics.add_warning(TemplateParseError::expression(
-                    "remote_branches() is deprecated; use remote_bookmarks() instead",
-                    function.name_span,
-                ));
-            }
+        |language, _diagnostics, _build_ctx, self_property, function| {
             function.expect_no_arguments()?;
             let index = language
                 .keyword_cache
@@ -955,11 +934,6 @@ fn builtin_commit_methods<'repo>() -> CommitTemplateBuildMethodFnMap<'repo, Comm
             Ok(L::wrap_ref_name_list(out_property))
         },
     );
-    // TODO: Remove the following block after jj 0.28+
-    map.insert("branches", map["bookmarks"]);
-    map.insert("local_branches", map["local_bookmarks"]);
-    map.insert("remote_branches", map["remote_bookmarks"]);
-
     map.insert(
         "tags",
         |language, _diagnostics, _build_ctx, self_property, function| {